♻️ history detail 데이터 불러오기 오타 수정 및 코드 작성
