+++ src/main/java/kakao/school/what/service/HistoryPopService.java
@@ -1,11 +1,9 @@
 package kakao.school.what.service;
 
 import kakao.school.what.domain.History;
-import kakao.school.what.domain.HistoryDetail;
 import kakao.school.what.dto.HistoryPopDTO;
-import kakao.school.what.repository.HistoryDetailPopRepository;
+import kakao.school.what.repository.HistoryDetailRepository;
 import kakao.school.what.repository.HistoryRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -13,7 +11,7 @@ public class HistoryPopService {
     private final HistoryRepository historyRepository;
 
 
-    public HistoryPopService(HistoryRepository historyRepository, HistoryDetailPopRepository historyDetailPopRepository) {
+    public HistoryPopService(HistoryRepository historyRepository, HistoryDetailRepository historyDetailRepository) {
         this.historyRepository = historyRepository;
 
     }
